# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TAOX API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/stocks/": {
            "get": {
                "tags": [
                    "stocks"
                ],
                "summary": "Get All Stocks",
                "description": "Obtiene todos los stocks químicos de una empresa",
                "operationId": "get_all_stocks_stocks__get",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChemicalStockResponseDTO"
                                    },
                                    "title": "Response Get All Stocks Stocks  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "stocks"
                ],
                "summary": "Create Stock",
                "description": "Crea un nuevo stock químico con umbrales de alertas automáticas",
                "operationId": "create_stock_stocks__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChemicalStockCreateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChemicalStockResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stocks/critical": {
            "get": {
                "tags": [
                    "stocks"
                ],
                "summary": "Get Critical Stocks",
                "description": "Obtiene todos los stocks en nivel crítico.\n\nDashboard para mostrar productos con stock crítico (cantidad_actual < cantidad_minima)",
                "operationId": "get_critical_stocks_stocks_critical_get",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChemicalStockResponseDTO"
                                    },
                                    "title": "Response Get Critical Stocks Stocks Critical Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stocks/{id_stock_quimico}": {
            "get": {
                "tags": [
                    "stocks"
                ],
                "summary": "Get Stock By Id",
                "description": "Obtiene un stock por ID con información de estado crítico y bajo",
                "operationId": "get_stock_by_id_stocks__id_stock_quimico__get",
                "parameters": [
                    {
                        "name": "id_stock_quimico",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Stock Quimico"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChemicalStockResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "stocks"
                ],
                "summary": "Update Stock",
                "description": "Actualiza un stock químico",
                "operationId": "update_stock_stocks__id_stock_quimico__put",
                "parameters": [
                    {
                        "name": "id_stock_quimico",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Stock Quimico"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChemicalStockUpdateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChemicalStockResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "stocks"
                ],
                "summary": "Delete Stock",
                "description": "Elimina un stock químico",
                "operationId": "delete_stock_stocks__id_stock_quimico__delete",
                "parameters": [
                    {
                        "name": "id_stock_quimico",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Stock Quimico"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChemicalStockCreateDTO": {
                "properties": {
                    "codigo_producto": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1,
                        "title": "Codigo Producto",
                        "description": "Código del producto químico"
                    },
                    "id_proceso": {
                        "type": "integer",
                        "title": "Id Proceso",
                        "description": "ID del proceso donde está el stock"
                    },
                    "cantidad_actual": {
                        "type": "number",
                        "minimum": 0.0,
                        "title": "Cantidad Actual",
                        "description": "Cantidad actual disponible"
                    },
                    "cantidad_minima": {
                        "type": "number",
                        "exclusiveMinimum": 0.0,
                        "title": "Cantidad Minima",
                        "description": "Cantidad mínima para alertas críticas"
                    },
                    "cantidad_maxima": {
                        "type": "number",
                        "exclusiveMinimum": 0.0,
                        "title": "Cantidad Maxima",
                        "description": "Capacidad máxima del proceso"
                    },
                    "unidad_medida": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 1,
                        "title": "Unidad Medida",
                        "description": "Unidad de medida (ml, g, l, kg, etc)"
                    },
                    "id_empresa": {
                        "type": "string",
                        "title": "Id Empresa",
                        "description": "ID de la empresa propietaria"
                    }
                },
                "type": "object",
                "required": [
                    "codigo_producto",
                    "id_proceso",
                    "cantidad_actual",
                    "cantidad_minima",
                    "cantidad_maxima",
                    "unidad_medida",
                    "id_empresa"
                ],
                "title": "ChemicalStockCreateDTO",
                "description": "DTO para crear un nuevo stock químico",
                "example": {
                    "cantidad_actual": 50.0,
                    "cantidad_maxima": 100.0,
                    "cantidad_minima": 10.0,
                    "codigo_producto": "CHEM001",
                    "id_empresa": "company-123",
                    "id_proceso": 1,
                    "unidad_medida": "ml"
                }
            },
            "ChemicalStockResponseDTO": {
                "properties": {
                    "id_stock_quimico": {
                        "type": "integer",
                        "title": "Id Stock Quimico"
                    },
                    "codigo_producto": {
                        "type": "string",
                        "title": "Codigo Producto"
                    },
                    "id_proceso": {
                        "type": "integer",
                        "title": "Id Proceso"
                    },
                    "cantidad_actual": {
                        "type": "number",
                        "title": "Cantidad Actual"
                    },
                    "cantidad_minima": {
                        "type": "number",
                        "title": "Cantidad Minima"
                    },
                    "cantidad_maxima": {
                        "type": "number",
                        "title": "Cantidad Maxima"
                    },
                    "unidad_medida": {
                        "type": "string",
                        "title": "Unidad Medida"
                    },
                    "id_empresa": {
                        "type": "string",
                        "title": "Id Empresa"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "es_stock_critico": {
                        "type": "boolean",
                        "title": "Es Stock Critico"
                    },
                    "es_stock_bajo": {
                        "type": "boolean",
                        "title": "Es Stock Bajo"
                    },
                    "porcentaje_stock": {
                        "type": "number",
                        "title": "Porcentaje Stock"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id_stock_quimico",
                    "codigo_producto",
                    "id_proceso",
                    "cantidad_actual",
                    "cantidad_minima",
                    "cantidad_maxima",
                    "unidad_medida",
                    "id_empresa",
                    "is_active",
                    "es_stock_critico",
                    "es_stock_bajo",
                    "porcentaje_stock",
                    "created_at",
                    "updated_at"
                ],
                "title": "ChemicalStockResponseDTO",
                "description": "DTO para respuestas de stock químico",
                "example": {
                    "cantidad_actual": 50.0,
                    "cantidad_maxima": 100.0,
                    "cantidad_minima": 10.0,
                    "codigo_producto": "CHEM001",
                    "created_at": "2025-03-01T10:00:00",
                    "es_stock_bajo": false,
                    "es_stock_critico": false,
                    "id_empresa": "company-123",
                    "id_proceso": 1,
                    "id_stock_quimico": 1,
                    "is_active": true,
                    "porcentaje_stock": 50.0,
                    "unidad_medida": "ml",
                    "updated_at": "2025-03-01T10:00:00"
                }
            },
            "ChemicalStockUpdateDTO": {
                "properties": {
                    "cantidad_actual": {
                        "anyOf": [
                            {
                                "type": "number",
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cantidad Actual",
                        "description": "Cantidad actual"
                    },
                    "cantidad_minima": {
                        "anyOf": [
                            {
                                "type": "number",
                                "exclusiveMinimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cantidad Minima",
                        "description": "Cantidad mínima"
                    },
                    "cantidad_maxima": {
                        "anyOf": [
                            {
                                "type": "number",
                                "exclusiveMinimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cantidad Maxima",
                        "description": "Cantidad máxima"
                    },
                    "unidad_medida": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Unidad Medida",
                        "description": "Unidad de medida"
                    },
                    "is_active": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Active",
                        "description": "Estado del stock"
                    }
                },
                "type": "object",
                "title": "ChemicalStockUpdateDTO",
                "description": "DTO para actualizar un stock químico",
                "example": {
                    "cantidad_actual": 45.0,
                    "cantidad_maxima": 110.0,
                    "cantidad_minima": 15.0
                }
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}